Incorporated new equalization options to make reverb eq process more robust.
Updated the demo script to showcase usage.

Now user can specify two new fields in BRIR_data:
- BRIR_data.BandsPerOctave: Specifies the frequency resolution for the
reverb equalization. The values can be 1 or 3. For smaller rooms, 1 is
recommended. By default it is 3.
- BRIR_data.EqTxx: To perform reverb equalization, the RT60 must be
estimated. However, there is generally not enough SNR to obtain a true
T60 estimation. This parameter allows the user to specify the desired
Txx for the RT60 estimation. Generally, a value of 30 is recommended.
However, this can blow up the RT estimation in cases where SNR is low.
For small rooms and very dry spaces we recommend using a value of 20. By
default it is 30.